pyproject.toml
[tool.poetry]
name = "fastapi-hotels"
version = "0.1.0"
description = ""
authors = ["arka-windows11 <arkad.sh@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
